:truck: move files into a components
