+++ src/components/CardInfo.js
