+++ src/components/InfoContainer.js
